feat!: added ConcreteTrigger objects to .execute()
This commit is a major architectural change to how message handling is handled.
It is UNTESTED.
The idea is to, for all commands, determine the actual triggering content of the message.
This is useful, since functions can now specify why they run, and the
commands can themselves make use of the RegExp matches/function matches and their indices.

- ICommand is now Command, an abstract class
- implemented Command#checkTrigggers(message, keyword, args)
- heavily refractored message.handler.ts
- updated ping.command.ts to avoid breaking changes