feat: added help command
Added the Help Command and integrated it into the 'Command List'.
It has to be integrated in a special way, due to it registering
every command inside the list on initialization! It has some
sub-commands, which might get extracted in the future. An idea
would be to make it another property `string => function`
inside the command class!