feat: added subcommands
Added Subcommands - a new Commands Mechanic, which will enable
you to write `command [sub-command]` and assign a function for
that combination. A default `help`-subcommand is always added,
which prints the syntax of the specified command!
Changes have been made to the Command Constructor, the
`checkTrigger`-function, which now searches recursively for
sub-commands and the changes have been applied to the ping
command and the default behaviour.
I decided to push this change, since sub-commands are a
beautiful way of both writing - as a user and categorizing
commands - as a developer. That way, for example all config
related commands can now be represented using commands like:
- `config get [name]`
- `config set [name] [value]`
- `config list`
- ...