fix: Fix of the Mess Canter has created
Just Kidding, we blame VCS. I also fixed a few other bugs, which weren't
related to the VCS merge, such as the fact, that as long as the prefix
was inside the message, no matter at which position, it was regarded as
a command! Also, I added the Subcommands Syntax to the Help Command.
You'll probably see when you go through it. Basically - everything's
working again!