feat: Added Sudo Command
Sudo Command will execute any command with Admin Privileges as
long as an admin executes it (to do: Permissions Service)! To
add a `sudo` command, you simply gotta add it into the sudo
subcommand list and a sudo handler in the command (if it can be
executed w/o the sudo command i.e. - it is directly accessible
through `commands/index.ts`)!